Use "saw credits" rather than "abandoned" for games.
Most games are abandoned, so the interesting outlier there is whether we
actually made it to the credits. Also show an icon for that (& allow sneaky
filter on "finished").

Also reorged item hierarchy a bit, make everything Consume that can be now that
Protracted is not properly shared by anything. Build can be started but
wouldn't be abandoned (though technically it could I guess...), books can be
started and abandoned, and games can be started and credits-ed.